Extract default implementation of DomainObjectIdentifier
We will move toward this general implementation instead of using several
parallel implementation.
